+++ docs/examples/UserGuide/settings.jl
@@ -1,19 +1,23 @@
-# TidierData comes with  the ability to log changes to the dataframe as operations/transformations are chained together, in addition to the ability to print the DataFrames.jl code underlying the many TidierData macros.
-# This page will review the `log` and `code` settings using the movies dataset.
+# TidierData.jl comes with two settings that make it easier to understand the transformations that are being applied to a data frame and to troubleshoot errors. These settings are `log` and `code`. The `log` setting outputs information about the data frame after each transformation, including the number of missing values and the number of unique values in each column. The `code` setting outputs the code that is being executed by the TidierData.jl macros. By default, both settings are set to `false`. This page will review the `log` and `code` settings using the movies dataset.
+# 
+# We recommend setting the `log` setting to `true` in general, and especially when you are first learning TidierData.jl. This will help you understand how the data frame is being transformed at each step. The `code` setting is useful for debugging errors in TidierData.jl chains.
 
 using TidierData
 using RDatasets
 
 movies = dataset("ggplot2", "movies");
 
-# ## log
-# Logging is set to `false` by default, but is toggled on and off as follows:
+# ## `log`
+# Logging is set to `false` by default but can enabled as follows:
+
 TidierData_set("log", true)
-# When enabled, each macro called will show information regarding each subsequent transformation of the data. 
-# Logging can be especially useful to catch silent bugs. 
-# When column values are changed, it will report how many new missing values there, the percent missing, and the number of unique values in that column as well. 
+
+# When enabled, each macro called will show information about its transformation of the data. Logging can be especially useful to catch silent bugs (those that do not result in an error). 
+# 
+# When column values are changed, it will report the number new missing values, the percentage of missing values, and the number of unique values.
+
 @chain movies begin 
-    @filter (Year > 2000)
+    @filter(Year > 2000)
     @mutate(Budget_cat = case_when(Budget > 18000 => "high",
                                    Budget > 2000  => "medium",
                                    Budget > 100 => "low",
@@ -22,14 +26,19 @@ TidierData_set("log", true)
     @group_by(Year, Budget_cat)
     @summarize(Avg_Budget = mean(Budget), n = n())
     @ungroup
-    @arrange n
+    @arrange(n)
 end
 
-# ## code
-# The code printing ability is also set to `false` by default. Its primary value would be for debugging errors within Tidier chains
-TidierData_set("code", true)
+TidierData_set("log", false) # disable logging
+
+# ## `code`
+# Code printing is set to `false` by default. Enabling this setting prints the underlying DataFrames.jl code created by TidierData.jl macros. It can be useful for debugging, especially for users who understand DataFrames.jl syntax, or for filing bug reports.
+
+TidierData_set("code", true) # enable macro code output
 
 @chain movies begin 
     @select(Title, Year, Budget)
-    @slice_sample n = 10
+    @slice_sample(n = 10)
 end
+
+TidierData_set("code", false) # disable macro code output